added cart functionality

+++ database/migrations/2023_04_03_090157_create_admins_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('email');
-            $table->integer('mobile');
+            $table->unsignedBigInteger('mobile');
             $table->string('password');
             $table->timestamps();
         });
